Move findRepoRoot functions from core to CLI workspace
Relocate repository discovery utilities from src/core/src/fs/ to src/cli/src/shared/
because they are CLI/test infrastructure, not core parsing functionality.

Original location: src/core/src/fs/find-repo-root.ts and find-repo-root-sync.ts
New location: src/cli/src/shared/find-repo-root.ts and find-repo-root-sync.ts

Rationale:
- These functions search for AGENTS.md and .git directories, which are development
  repository sentinels, not GameMaker project files
- Only used by CLI commands and test infrastructure, never by parser or plugin
- findRepoRoot/findRepoRootSync were the only FS utilities in Core that were
  specific to this repository's structure rather than general filesystem operations
- Keeping them in Core violated separation of concerns (parser/formatter vs tooling)

Changes:
- Move both function implementations to CLI shared directory
- Update imports to use Core.walkAncestorDirectories (which stays in Core as it's
  a general directory traversal utility used by semantic workspace)
- Move test from src/core/test/fs/ to src/cli/test/
- Remove exports from Core.fs and Core namespace
- Add exports to CLI.Shared namespace
- Update all consumers (CLI tests, runtime-wrapper test) to import from CLI.Shared
- Add @gml-modules/cli dependency to runtime-wrapper package.json
- All tests pass after the move
