+++ src/cli/src/shared/index.ts
@@ -3,6 +3,8 @@
 // while providing access to the upstream Core namespace.
 export { Core } from "@gml-modules/core";
 
+export * from "./find-repo-root.js";
+export * from "./find-repo-root-sync.js";
 export * from "./module.js";
 export * from "./workspace-paths.js";
 export * from "./debounce.js";
