+++ src/cli/src/shared/workspace-paths.ts
@@ -1,7 +1,8 @@
 import fs from "node:fs";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
-import { Core } from "@gml-modules/core";
+
+import { findRepoRootSync } from "./find-repo-root-sync.js";
 
 /**
  * Centralized workspace path helpers for the CLI package. Several modules
@@ -50,7 +51,7 @@ function resolveCliPackageDirectory(startDirectory: string): string {
 const CLI_PACKAGE_DIRECTORY = resolveCliPackageDirectory(SHARED_DIRECTORY);
 const CLI_SRC_DIRECTORY = path.resolve(CLI_PACKAGE_DIRECTORY, "src");
 const WORKSPACE_SOURCE_DIRECTORY = path.resolve(CLI_PACKAGE_DIRECTORY, "..");
-const REPO_ROOT = Core.findRepoRootSync(SHARED_DIRECTORY);
+const REPO_ROOT = findRepoRootSync(SHARED_DIRECTORY);
 
 export {
     CLI_SRC_DIRECTORY,
