+++ src/core/src/fs/index.ts
@@ -1,4 +1,2 @@
 export * from "./path.js";
 export * from "./io.js";
-export * from "./find-repo-root.js";
-export * from "./find-repo-root-sync.js";
