+++ src/runtime-wrapper/package.json
@@ -12,6 +12,7 @@
     "dist"
   ],
   "dependencies": {
+    "@gml-modules/cli": "0.0.1",
     "@gml-modules/core": "0.0.1"
   },
   "scripts": {
